+++ spec/rails/form/check_box_helper_spec.rb
@@ -23,6 +23,15 @@ describe 'check_box' do
       it { expect(form).to include 'Given name</label>' }
     end
 
+    context 'not given a help option, does not display a help box' do
+      it { expect(form).not_to include 'help-block' }
+    end
+
+    context 'given a help option, displays a help box' do
+      let(:options) { {help: 'Please select an option'} }
+      it { expect(form).to include '<span class="help-block text-left">Please select an option</span>' }
+    end
+
     specify 'not given an error, does not apply has-error to the form group' do
       expect(form).not_to include 'has-error'
     end
@@ -80,6 +89,14 @@ describe 'check_box' do
 
       specify 'applies sr-only to the error message' do
         expect(form).to include '<span class="help-block text-left sr-only">cannot be nil</span>'
+      end
+    end
+
+    context 'given a help message' do
+      let(:options) { {help: 'Please select an option'} }
+
+      specify 'applies sr-only to the help message' do
+        expect(form).to include '<span class="help-block text-left sr-only">Please select an option</span>'
       end
     end