+++ CHANGELOG.md
@@ -8,6 +8,7 @@ For more information about changelogs, check
 
 ## 1.3.0 - unreleased
 
+* [ENHANCEMENT] Add `:help` option to display a help block after most form fields
 * [FEATURE] Extend `button_to` to be wrapped in <li> when inside a `nav`, just like `link_to`
 * [ENHANCEMENT] Bump Bootstrap version to 3.3.2
 * [ENHANCEMENT] Bump Font Awesome version to 4.3.0